Rename adapter_app to test_app to handle future changes
